\name{LendingClubJFS-package}
\alias{LendingClubJFS-package}
\alias{LendingClubJFS}
\docType{package}
\title{
\packageTitle{LendingClubJFS}
}
\description{
\packageDescription{LendingClubJFS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LendingClubJFS}
\packageIndices{LendingClubJFS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LendingClubJFS}

Maintainer: \packageMaintainer{LendingClubJFS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
